requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
django-filter
psycopg2-binary
celery[redis]
redis
pillow
pytesseract
torch
transformers
scikit-learn
pandas
prophet
gunicorn
python-dotenv
cloudinary
django-cloudinary-storage
spacy
python-dateutil
nepali-datetime
